ENH: Refactor issues to pages script
Refactor issues to pages script: split the existing script into a module
to host the methods, and the actual script that makes use of such
methods.

Paves the way towards an easier maintenance.
