+++ .github/workflows/issue-to-page.yml
@@ -26,7 +26,7 @@ jobs:
       run: |
         url="https://${GITHUB_API_URL}/repos/${GITHUB_REPOSITORY}/issues?per_page=100"
         path="content/project"
-        python scripts/issues_to_pages.py $url $path
+        python scripts/transform_issues_to_pages.py $url $path
 
     - name: Create Pull Request
       uses: peter-evans/create-pull-request@v3
